#2339 - Moved document objects to separate directory (let PHP < 8 tests skip but not fail)
